moved the test package to the pkg dir so that it can be imported by extensions if needed
