--- pkg/mime/mime.go
@@ -1,693 +1,0 @@
-package mime
-
-type Type string
-
-const (
-	// Application Types
-	ApplicationAndrewInset                               Type = "application/andrew-inset"
-	ApplicationApplixware                                Type = "application/applixware"
-	ApplicationAtomXML                                   Type = "application/atom+xml"
-	ApplicationAtomcatXML                                Type = "application/atomcat+xml"
-	ApplicationAtomsvcXML                                Type = "application/atomsvc+xml"
-	ApplicationCCXML                                     Type = "application/ccxml+xml"
-	ApplicationCuSeeme                                   Type = "application/cu-seeme"
-	ApplicationDavmountXML                               Type = "application/davmount+xml"
-	ApplicationEcmascript                                Type = "application/ecmascript"
-	ApplicationEmmaXML                                   Type = "application/emma+xml"
-	ApplicationEpubZip                                   Type = "application/epub+zip"
-	ApplicationFontTDPFR                                 Type = "application/font-tdpfr"
-	ApplicationGzip                                      Type = "application/gzip"
-	ApplicationHyperstudio                               Type = "application/hyperstudio"
-	ApplicationJavaArchive                               Type = "application/java-archive"
-	ApplicationJavaSerializedObject                      Type = "application/java-serialized-object"
-	ApplicationJavaVM                                    Type = "application/java-vm"
-	ApplicationJSON                                      Type = "application/json"
-	ApplicationLostXML                                   Type = "application/lost+xml"
-	ApplicationMacBinHex40                               Type = "application/mac-binhex40"
-	ApplicationMacCompactPro                             Type = "application/mac-compactpro"
-	ApplicationMarc                                      Type = "application/marc"
-	ApplicationMathematica                               Type = "application/mathematica"
-	ApplicationMathMLXML                                 Type = "application/mathml+xml"
-	ApplicationMbox                                      Type = "application/mbox"
-	ApplicationMediaServerControlXML                     Type = "application/mediaservercontrol+xml"
-	ApplicationMp4                                       Type = "application/mp4"
-	ApplicationMSWord                                    Type = "application/msword"
-	ApplicationMXF                                       Type = "application/mxf"
-	ApplicationOctetStream                               Type = "application/octet-stream"
-	ApplicationODA                                       Type = "application/oda"
-	ApplicationOebpsPackageXML                           Type = "application/oebps-package+xml"
-	ApplicationOgg                                       Type = "application/ogg"
-	ApplicationOneNote                                   Type = "application/onenote"
-	ApplicationPatchOpsErrorXML                          Type = "application/patch-ops-error+xml"
-	ApplicationPDF                                       Type = "application/pdf"
-	ApplicationPGPEncrypted                              Type = "application/pgp-encrypted"
-	ApplicationPGPSignature                              Type = "application/pgp-signature"
-	ApplicationPicsRules                                 Type = "application/pics-rules"
-	ApplicationPKCS10                                    Type = "application/pkcs10"
-	ApplicationPKCS7Mime                                 Type = "application/pkcs7-mime"
-	ApplicationPKCS7Signature                            Type = "application/pkcs7-signature"
-	ApplicationPKIXCert                                  Type = "application/pkix-cert"
-	ApplicationPKIXCRL                                   Type = "application/pkix-crl"
-	ApplicationPKIXPkiPath                               Type = "application/pkix-pkipath"
-	ApplicationPKIXCMP                                   Type = "application/pkixcmp"
-	ApplicationPLSXML                                    Type = "application/pls+xml"
-	ApplicationPostscript                                Type = "application/postscript"
-	ApplicationPRQL                                      Type = "application/prql"
-	ApplicationPRS_CWW                                   Type = "application/prs.cww"
-	ApplicationRDFXML                                    Type = "application/rdf+xml"
-	ApplicationRegInfoXML                                Type = "application/reginfo+xml"
-	ApplicationRelaxNGCompactSyntax                      Type = "application/relax-ng-compact-syntax"
-	ApplicationResourceListsXML                          Type = "application/resource-lists+xml"
-	ApplicationResourceListsDiffXML                      Type = "application/resource-lists-diff+xml"
-	ApplicationRLSServicesXML                            Type = "application/rls-services+xml"
-	ApplicationRSDXML                                    Type = "application/rsd+xml"
-	ApplicationRSSXML                                    Type = "application/rss+xml"
-	ApplicationRTF                                       Type = "application/rtf"
-	ApplicationSBMLXML                                   Type = "application/sbml+xml"
-	ApplicationSCVP_CVRequest                            Type = "application/scvp-cv-request"
-	ApplicationSCVP_CVResponse                           Type = "application/scvp-cv-response"
-	ApplicationSCVP_VPRequest                            Type = "application/scvp-vp-request"
-	ApplicationSCVP_VPResponse                           Type = "application/scvp-vp-response"
-	ApplicationSDP                                       Type = "application/sdp"
-	ApplicationSetPaymentInitiation                      Type = "application/set-payment-initiation"
-	ApplicationSetRegistrationInitiation                 Type = "application/set-registration-initiation"
-	ApplicationSHFXML                                    Type = "application/shf+xml"
-	ApplicationSMILXML                                   Type = "application/smil+xml"
-	ApplicationSPARQLQuery                               Type = "application/sparql-query"
-	ApplicationSPARQLResultsXML                          Type = "application/sparql-results+xml"
-	ApplicationSRGS                                      Type = "application/srgs"
-	ApplicationSRGSXML                                   Type = "application/srgs+xml"
-	ApplicationSSMLXML                                   Type = "application/ssml+xml"
-	ApplicationVND3GPPPicBWLarge                         Type = "application/vnd.3gpp.pic-bw-large"
-	ApplicationVND3GPPPicBWSmall                         Type = "application/vnd.3gpp.pic-bw-small"
-	ApplicationVND3GPPPicBWVar                           Type = "application/vnd.3gpp.pic-bw-var"
-	ApplicationVND3GPP2Tcap                              Type = "application/vnd.3gpp2.tcap"
-	ApplicationVND3MPostItNotes                          Type = "application/vnd.3m.post-it-notes"
-	ApplicationVNDAccpacSimplyASO                        Type = "application/vnd.accpac.simply.aso"
-	ApplicationVNDAccpacSimplyIMP                        Type = "application/vnd.accpac.simply.imp"
-	ApplicationVNDAcucobol                               Type = "application/vnd.acucobol"
-	ApplicationVNDAcucorp                                Type = "application/vnd.acucorp"
-	ApplicationVNDAirInstallerPackage                    Type = "application/vnd.adobe.air-application-installer-package+zip"
-	ApplicationVNDAdobeXDPXML                            Type = "application/vnd.adobe.xdp+xml"
-	ApplicationVNDAdobeXFDF                              Type = "application/vnd.adobe.xfdf"
-	ApplicationVNDAirzipFileSecureAZF                    Type = "application/vnd.airzip.filesecure.azf"
-	ApplicationVNDAirzipFileSecureAZS                    Type = "application/vnd.airzip.filesecure.azs"
-	ApplicationVNDAmazonEbook                            Type = "application/vnd.amazon.ebook"
-	ApplicationVNDAmericanDynamicsAcc                    Type = "application/vnd.americandynamics.acc"
-	ApplicationVNDAmigaAMI                               Type = "application/vnd.amiga.ami"
-	ApplicationVNDCAPackageArchive                       Type = "application/vnd.android.package-archive"
-	ApplicationVNDAnserWebCertificateIssueInitiation     Type = "application/vnd.anser-web-certificate-issue-initiation"
-	ApplicationVNDAnserWebFundsTransferInitiation        Type = "application/vnd.anser-web-funds-transfer-initiation"
-	ApplicationVNDAntixGameComponent                     Type = "application/vnd.antix.game-component"
-	ApplicationVNDAppleInstaller                         Type = "application/vnd.apple.installer+xml"
-	ApplicationVNDArastraSWI                             Type = "application/vnd.arastra.swi"
-	ApplicationVNDAudiograph                             Type = "application/vnd.audiograph"
-	ApplicationVNDBlueiceMultipass                       Type = "application/vnd.blueice.multipass"
-	ApplicationVNDBMI                                    Type = "application/vnd.bmi"
-	ApplicationVNDBusinessobjects                        Type = "application/vnd.businessobjects"
-	ApplicationVNDChemDrawXML                            Type = "application/vnd.chemdraw+xml"
-	ApplicationVNDChipnutsKaraokeMMD                     Type = "application/vnd.chipnuts.karaoke-mmd"
-	ApplicationVNDCinderella                             Type = "application/vnd.cinderella"
-	ApplicationVNDClaymore                               Type = "application/vnd.claymore"
-	ApplicationVNDClonkC4Group                           Type = "application/vnd.clonk.c4group"
-	ApplicationVNDCommonspace                            Type = "application/vnd.commonspace"
-	ApplicationVNDContactCMSG                            Type = "application/vnd.contact.cmsg"
-	ApplicationVNDCosmocaller                            Type = "application/vnd.cosmocaller"
-	ApplicationVNDCrickClicker                           Type = "application/vnd.crick.clicker"
-	ApplicationVNDCrickClickerKeyboard                   Type = "application/vnd.crick.clicker.keyboard"
-	ApplicationVNDCrickClickerPalette                    Type = "application/vnd.crick.clicker.palette"
-	ApplicationVNDCrickClickerTemplate                   Type = "application/vnd.crick.clicker.template"
-	ApplicationVNDCrickClickerWordbank                   Type = "application/vnd.crick.clicker.wordbank"
-	ApplicationVNDCriticalToolsWBSXML                    Type = "application/vnd.criticaltools.wbs+xml"
-	ApplicationVNDCTCPosml                               Type = "application/vnd.ctc-posml"
-	ApplicationVNDCUPS_PPD                               Type = "application/vnd.cups-ppd"
-	ApplicationVNDCurlCar                                Type = "application/vnd.curl.car"
-	ApplicationVNDCurlPCurl                              Type = "application/vnd.curl.pcurl"
-	ApplicationVNDDataVisionRDZ                          Type = "application/vnd.data-vision.rdz"
-	ApplicationVNDDebianBinaryPackage                    Type = "application/vnd.debian.binary-package"
-	ApplicationVNDDenovoFCSELAYOUTLINK                   Type = "application/vnd.denovo.fcselayout-link"
-	ApplicationVNDNA                                     Type = "application/vnd.dna"
-	ApplicationVNDDolbyMLP                               Type = "application/vnd.dolby.mlp"
-	ApplicationVNDDPGraph                                Type = "application/vnd.dpgraph"
-	ApplicationVNDDreamFactory                           Type = "application/vnd.dreamfactory"
-	ApplicationVNDDynageo                                Type = "application/vnd.dynageo"
-	ApplicationVNDEcowinChart                            Type = "application/vnd.ecowin.chart"
-	ApplicationVNDEnliven                                Type = "application/vnd.enliven"
-	ApplicationVNDEpsonESF                               Type = "application/vnd.epson.esf"
-	ApplicationVNDEpsonMSF                               Type = "application/vnd.epson.msf"
-	ApplicationVNDEpsonQuickAnime                        Type = "application/vnd.epson.quickanime"
-	ApplicationVNDEpsonSalt                              Type = "application/vnd.epson.salt"
-	ApplicationVNDEpsonSSF                               Type = "application/vnd.epson.ssf"
-	ApplicationVNDExzigno3XML                            Type = "application/vnd.eszigno3+xml"
-	ApplicationVNDEzpixAlbum                             Type = "application/vnd.ezpix-album"
-	ApplicationVNDEzpixPackage                           Type = "application/vnd.ezpix-package"
-	ApplicationVNDFDF                                    Type = "application/vnd.fdf"
-	ApplicationVNDFDSN_MSEED                             Type = "application/vnd.fdsn.mseed"
-	ApplicationVNDFDSN_Seed                              Type = "application/vnd.fdsn.seed"
-	ApplicationVNDFlographit                             Type = "application/vnd.flographit"
-	ApplicationVNDFluxtimeClip                           Type = "application/vnd.fluxtime.clip"
-	ApplicationVNDFrameMaker                             Type = "application/vnd.framemaker"
-	ApplicationVNDFrogansFNC                             Type = "application/vnd.frogans.fnc"
-	ApplicationVNDFrogansLTF                             Type = "application/vnd.frogans.ltf"
-	ApplicationVNDFSCWeblaunch                           Type = "application/vnd.fsc.weblaunch"
-	ApplicationVNDFujitsuOASYS                           Type = "application/vnd.fujitsu.oasys"
-	ApplicationVNDFAFujitsuOASYS2                        Type = "application/vnd.fujitsu.oasys2"
-	ApplicationVNDFAFujitsuOASYS3                        Type = "application/vnd.fujitsu.oasys3"
-	ApplicationVNDFAFujitsuOASYSGP                       Type = "application/vnd.fujitsu.oasysgp"
-	ApplicationVNDFAFujitsuOASYSPRS                      Type = "application/vnd.fujitsu.oasysprs"
-	ApplicationVNDFFujixeroxDDD                          Type = "application/vnd.fujixerox.ddd"
-	ApplicationVNDFFujixeroxDocuworks                    Type = "application/vnd.fujixerox.docuworks"
-	ApplicationVNDFFujixeroxDocuworksBinder              Type = "application/vnd.fujixerox.docuworks.binder"
-	ApplicationVNDFuzzySheet                             Type = "application/vnd.fuzzysheet"
-	ApplicationVNDGenomatixTuxedo                        Type = "application/vnd.genomatix.tuxedo"
-	ApplicationVNDGeogebraFile                           Type = "application/vnd.geogebra.file"
-	ApplicationVNDGeogebraTool                           Type = "application/vnd.geogebra.tool"
-	ApplicationVNDGeometryExplorer                       Type = "application/vnd.geometry-explorer"
-	ApplicationVNDGerber                                 Type = "application/vnd.gerber"
-	ApplicationVNDGMX                                    Type = "application/vnd.gmx"
-	ApplicationVNDGoogleEarthKMLXML                      Type = "application/vnd.google-earth.kml+xml"
-	ApplicationVNDGoogleEarthKMZ                         Type = "application/vnd.google-earth.kmz"
-	ApplicationVNDGrafeq                                 Type = "application/vnd.grafeq"
-	ApplicationVNDGrooveAccount                          Type = "application/vnd.groove-account"
-	ApplicationVNDGrooveHelp                             Type = "application/vnd.groove-help"
-	ApplicationVNDGrooveIdentityMessage                  Type = "application/vnd.groove-identity-message"
-	ApplicationVNDGrooveInjector                         Type = "application/vnd.groove-injector"
-	ApplicationVNDGrooveToolMessage                      Type = "application/vnd.groove-tool-message"
-	ApplicationVNDGrooveToolTemplate                     Type = "application/vnd.groove-tool-template"
-	ApplicationVNDGrooveVCard                            Type = "application/vnd.groove-vcard"
-	ApplicationVNDHandheldEntertainmentXML               Type = "application/vnd.handheld-entertainment+xml"
-	ApplicationVNDHBCI                                   Type = "application/vnd.hbci"
-	ApplicationVNDHHELessonPlayer                        Type = "application/vnd.hhe.lesson-player"
-	ApplicationVNDHPHPGL                                 Type = "application/vnd.hp-hpgl"
-	ApplicationVNDHPHPID                                 Type = "application/vnd.hp-hpid"
-	ApplicationVNDHPHPS                                  Type = "application/vnd.hp-hps"
-	ApplicationVNDHPJLYT                                 Type = "application/vnd.hp-jlyt"
-	ApplicationVNDHPPCL                                  Type = "application/vnd.hp-pcl"
-	ApplicationVNDHPPCLXL                                Type = "application/vnd.hp-pclxl"
-	ApplicationVNDHydrostatixSofData                     Type = "application/vnd.hydrostatix.sof-data"
-	ApplicationVNDHZN3DCrossword                         Type = "application/vnd.hzn-3d-crossword"
-	ApplicationVNDIBMMinipay                             Type = "application/vnd.ibm.minipay"
-	ApplicationVNDIBMModcap                              Type = "application/vnd.ibm.modcap"
-	ApplicationVNDIBMRightsManagement                    Type = "application/vnd.ibm.rights-management"
-	ApplicationVNDIBMContainer                           Type = "application/vnd.ibm.secure-container"
-	ApplicationVNDICCProfile                             Type = "application/vnd.iccprofile"
-	ApplicationVNDIGLoader                               Type = "application/vnd.igloader"
-	ApplicationVNDImmervisionIVP                         Type = "application/vnd.immervision-ivp"
-	ApplicationVNDImmervisionIVU                         Type = "application/vnd.immervision-ivu"
-	ApplicationVNDInterconFormnet                        Type = "application/vnd.intercon.formnet"
-	ApplicationVNDIntuQBO                                Type = "application/vnd.intu.qbo"
-	ApplicationVNDIntuQFX                                Type = "application/vnd.intu.qfx"
-	ApplicationVNDIPUnpluggedRCProfile                   Type = "application/vnd.ipunplugged.rcprofile"
-	ApplicationVNDIrepositoryPackageXML                  Type = "application/vnd.irepository.package+xml"
-	ApplicationVNDISXPR                                  Type = "application/vnd.is-xpr"
-	ApplicationVNDJam                                    Type = "application/vnd.jam"
-	ApplicationVNDJCPJavaME                              Type = "application/vnd.jcp.javame.midlet-rms"
-	ApplicationVNDJISP                                   Type = "application/vnd.jisp"
-	ApplicationVNDJoostJodaArchive                       Type = "application/vnd.joost.joda-archive"
-	ApplicationVNDKahootz                                Type = "application/vnd.kahootz"
-	ApplicationVNDKDEKarbon                              Type = "application/vnd.kde.karbon"
-	ApplicationVNDKDEKChart                              Type = "application/vnd.kde.kchart"
-	ApplicationVNDKDEKFormula                            Type = "application/vnd.kde.kformula"
-	ApplicationVNDKDEKivio                               Type = "application/vnd.kde.kivio"
-	ApplicationVNDKDEKontour                             Type = "application/vnd.kde.kontour"
-	ApplicationVNDKDEPp                                  Type = "application/vnd.kde.kpresenter"
-	ApplicationVNDKDESpread                              Type = "application/vnd.kde.kspread"
-	ApplicationVNDKDEKWord                               Type = "application/vnd.kde.kword"
-	ApplicationVNDKenameaapp                             Type = "application/vnd.kenameaapp"
-	ApplicationVNDKidspiration                           Type = "application/vnd.kidspiration"
-	ApplicationVNDKinar                                  Type = "application/vnd.kinar"
-	ApplicationVNDKoan                                   Type = "application/vnd.koan"
-	ApplicationVNDKodakDescriptor                        Type = "application/vnd.kodak-descriptor"
-	ApplicationVNDLlamaGraphicsLifeBalanceDesktop        Type = "application/vnd.llamagraphics.life-balance.desktop"
-	ApplicationVNDLlamaGraphicsLifeBalanceExchangeXML    Type = "application/vnd.llamagraphics.life-balance.exchange+xml"
-	ApplicationVNDLotus123                               Type = "application/vnd.lotus-1-2-3"
-	ApplicationVNDLotusApproach                          Type = "application/vnd.lotus-approach"
-	ApplicationVNDLotusFreelance                         Type = "application/vnd.lotus-freelance"
-	ApplicationVNDLotusNotes                             Type = "application/vnd.lotus-notes"
-	ApplicationVNDLotusOrganizer                         Type = "application/vnd.lotus-organizer"
-	ApplicationVNDLotusScreencam                         Type = "application/vnd.lotus-screencam"
-	ApplicationVNDLotusWordPro                           Type = "application/vnd.lotus-wordpro"
-	ApplicationVNDMacPortsPortPkg                        Type = "application/vnd.macports.portpkg"
-	ApplicationVNDCMCD                                   Type = "application/vnd.mcd"
-	ApplicationVNDMedCalcData                            Type = "application/vnd.medcalcdata"
-	ApplicationVNDMediaStationCDKey                      Type = "application/vnd.mediastation.cdkey"
-	ApplicationVNDMFER                                   Type = "application/vnd.mfer"
-	ApplicationVNDMFMP                                   Type = "application/vnd.mfmp"
-	ApplicationVNDMicrografxFlo                          Type = "application/vnd.micrografx.flo"
-	ApplicationVNDMicrografxIgx                          Type = "application/vnd.micrografx.igx"
-	ApplicationVNDMIF                                    Type = "application/vnd.mif"
-	ApplicationVNDMobiusDAF                              Type = "application/vnd.mobius.daf"
-	ApplicationVNDMobiusDIS                              Type = "application/vnd.mobius.dis"
-	ApplicationVNDMobiusMBK                              Type = "application/vnd.mobius.mbk"
-	ApplicationVNDMobiusMQY                              Type = "application/vnd.mobius.mqy"
-	ApplicationVNDMobiusMSL                              Type = "application/vnd.mobius.msl"
-	ApplicationVNDMobiusPLC                              Type = "application/vnd.mobius.plc"
-	ApplicationVNDMobiusTXF                              Type = "application/vnd.mobius.txf"
-	ApplicationVNDMophunApplication                      Type = "application/vnd.mophun.application"
-	ApplicationVNDMophunCertificate                      Type = "application/vnd.mophun.certificate"
-	ApplicationVNDMozillaXULXML                          Type = "application/vnd.mozilla.xul+xml"
-	ApplicationVNDMSArtGallery                           Type = "application/vnd.ms-artgalry"
-	ApplicationVNDMSCabCompressed                        Type = "application/vnd.ms-cab-compressed"
-	ApplicationVNDMSExcel                                Type = "application/vnd.ms-excel"
-	ApplicationVNDMSExcelAddinMacroEnabled12             Type = "application/vnd.ms-excel.addin.macroenabled.12"
-	ApplicationVNDMSExcelSheetBinaryMacroEnabled12       Type = "application/vnd.ms-excel.sheet.binary.macroenabled.12"
-	ApplicationVNDMSExcelSheetMacroEnabled12             Type = "application/vnd.ms-excel.sheet.macroenabled.12"
-	ApplicationVNDMSExcelTemplateMacroEnabled12          Type = "application/vnd.ms-excel.template.macroenabled.12"
-	ApplicationVNDMSFontObject                           Type = "application/vnd.ms-fontobject"
-	ApplicationVNDMSHtmlHelp                             Type = "application/vnd.ms-htmlhelp"
-	ApplicationVNDMSIMS                                  Type = "application/vnd.ms-ims"
-	ApplicationVNDMSLRM                                  Type = "application/vnd.ms-lrm"
-	ApplicationVNDMSPKISeccat                            Type = "application/vnd.ms-pki.seccat"
-	ApplicationVNDMSPKIStl                               Type = "application/vnd.ms-pki.stl"
-	ApplicationVNDMSPowerPoint                           Type = "application/vnd.ms-powerpoint"
-	ApplicationVNDMSPowerPointAddinMacroEnabled12        Type = "application/vnd.ms-powerpoint.addin.macroenabled.12"
-	ApplicationVNDMSPowerPointPresentationMacroEnabled12 Type = "application/vnd.ms-powerpoint.presentation.macroenabled.12"
-	ApplicationVNDMSPowerPointSlideMacroEnabled12        Type = "application/vnd.ms-powerpoint.slide.macroenabled.12"
-	ApplicationVNDMSPowerPointSlideshowMacroEnabled12    Type = "application/vnd.ms-powerpoint.slideshow.macroenabled.12"
-	ApplicationVNDMSPowerPointTemplateMacroEnabled12     Type = "application/vnd.ms-powerpoint.template.macroenabled.12"
-	ApplicationVNDMSProject                              Type = "application/vnd.ms-project"
-	ApplicationVNDMSWordDocumentMacroEnabled12           Type = "application/vnd.ms-word.document.macroenabled.12"
-	ApplicationVNDMSWordTemplateMacroEnabled12           Type = "application/vnd.ms-word.template.macroenabled.12"
-	ApplicationVNDMSWorks                                Type = "application/vnd.ms-works"
-	ApplicationVNDMSWPL                                  Type = "application/vnd.ms-wpl"
-	ApplicationVNDMSXPSDocument                          Type = "application/vnd.ms-xpsdocument"
-	ApplicationVNDMSeq                                   Type = "application/vnd.mseq"
-	ApplicationVNDMusician                               Type = "application/vnd.musician"
-	ApplicationVNDMuveeStyle                             Type = "application/vnd.muvee.style"
-	ApplicationVNDNeurolanguageNLU                       Type = "application/vnd.neurolanguage.nlu"
-	ApplicationVNDNoblenetDirectory                      Type = "application/vnd.noblenet-directory"
-	ApplicationVNDNoblenetSealer                         Type = "application/vnd.noblenet-sealer"
-	ApplicationVNDNoblenetWeb                            Type = "application/vnd.noblenet-web"
-	ApplicationVNDNokiaNGageData                         Type = "application/vnd.nokia.n-gage.data"
-	ApplicationVNDNokiaNGageSymbianInstall               Type = "application/vnd.nokia.n-gage.symbian.install"
-	ApplicationVNDNokiaRadioPreset                       Type = "application/vnd.nokia.radio-preset"
-	ApplicationVNDNokiaRadioPresets                      Type = "application/vnd.nokia.radio-presets"
-	ApplicationVNDNovadigmEDM                            Type = "application/vnd.novadigm.edm"
-	ApplicationVNDNovadigmEDX                            Type = "application/vnd.novadigm.edx"
-	ApplicationVNDNovadigmExt                            Type = "application/vnd.novadigm.ext"
-	ApplicationVNDOasisOpenDocumentChart                 Type = "application/vnd.oasis.opendocument.chart"
-	ApplicationVNDOasisOpenDocumentChartTemplate         Type = "application/vnd.oasis.opendocument.chart-template"
-	ApplicationVNDOasisOpenDocumentDatabase              Type = "application/vnd.oasis.opendocument.database"
-	ApplicationVNDOasisOpenDocumentFormula               Type = "application/vnd.oasis.opendocument.formula"
-	ApplicationVNDOasisOpenDocumentFormulaTemplate       Type = "application/vnd.oasis.opendocument.formula-template"
-	ApplicationVNDOasisOpenDocumentGraphics              Type = "application/vnd.oasis.opendocument.graphics"
-	ApplicationVNDOasisOpenDocumentGraphicsTemplate      Type = "application/vnd.oasis.opendocument.graphics-template"
-	ApplicationVNDOasisOpenDocumentImage                 Type = "application/vnd.oasis.opendocument.image"
-	ApplicationVNDOasisOpenDocumentImageTemplate         Type = "application/vnd.oasis.opendocument.image-template"
-	ApplicationVNDOasisOpenDocumentPresentation          Type = "application/vnd.oasis.opendocument.presentation"
-	ApplicationVNDOasisOpenDocumentPresentationTemplate  Type = "application/vnd.oasis.opendocument.presentation-template"
-	ApplicationVNDOasisOpenDocumentSpreadsheet           Type = "application/vnd.oasis.opendocument.spreadsheet"
-	ApplicationVNDOasisOpenDocumentSpreadsheetTemplate   Type = "application/vnd.oasis.opendocument.spreadsheet-template"
-	ApplicationVNDOasisOpenDocumentText                  Type = "application/vnd.oasis.opendocument.text"
-	ApplicationVNDOasisOpenDocumentTextMaster            Type = "application/vnd.oasis.opendocument.text-master"
-	ApplicationVNDOasisOpenDocumentTextTemplate          Type = "application/vnd.oasis.opendocument.text-template"
-	ApplicationVNDOasisOpenDocumentTextWeb               Type = "application/vnd.oasis.opendocument.text-web"
-	ApplicationVNDOLPCSugar                              Type = "application/vnd.olpc-sugar"
-	ApplicationVNDOMA_DD2XML                             Type = "application/vnd.oma.dd2+xml"
-	ApplicationVNDOpenOfficeOrgExtension                 Type = "application/vnd.openofficeorg.extension"
-	ApplicationVNDOpenXMLPresentationML                  Type = "application/vnd.openxmlformats-officedocument.presentationml.presentation"
-	ApplicationVNDOpenXMLPresentationMLSlide             Type = "application/vnd.openxmlformats-officedocument.presentationml.slide"
-	ApplicationVNDOpenXMLPresentationMLSlideshow         Type = "application/vnd.openxmlformats-officedocument.presentationml.slideshow"
-	ApplicationVNDOpenXMLPresentationMLTemplate          Type = "application/vnd.openxmlformats-officedocument.presentationml.template"
-	ApplicationVNDOpenXMLSpreadsheetML                   Type = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-	ApplicationVNDOpenXMLSpreadsheetMLTemplate           Type = "application/vnd.openxmlformats-officedocument.spreadsheetml.template"
-	ApplicationVNDOpenXMLWordProcessingMLDocument        Type = "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
-	ApplicationVNDOpenXMLWordProcessingMLTemplate        Type = "application/vnd.openxmlformats-officedocument.wordprocessingml.template"
-	ApplicationVNDOSGiDP                                 Type = "application/vnd.osgi.dp"
-	ApplicationVNDPalm                                   Type = "application/vnd.palm"
-	ApplicationVNDPGFormat                               Type = "application/vnd.pg.format"
-	ApplicationVNDPGOsasli                               Type = "application/vnd.pg.osasli"
-	ApplicationVNDPicsel                                 Type = "application/vnd.picsel"
-	ApplicationVNDPocketLearn                            Type = "application/vnd.pocketlearn"
-	ApplicationVNDPowerBuilder6                          Type = "application/vnd.powerbuilder6"
-	ApplicationVNDPPreviewSystemsBox                     Type = "application/vnd.previewsystems.box"
-	ApplicationVNDProteusMagazine                        Type = "application/vnd.proteus.magazine"
-	ApplicationVNDPublishareDeltaTree                    Type = "application/vnd.publishare-delta-tree"
-	ApplicationVNDPviPTID1                               Type = "application/vnd.pvi.ptid1"
-	ApplicationVNDQuarkQuarkXPress                       Type = "application/vnd.quark.quarkxpress"
-	ApplicationVNDRAR                                    Type = "application/vnd.rar"
-	ApplicationVNDRecordareMusicXML                      Type = "application/vnd.recordare.musicxml"
-	ApplicationVNDRecordareMusicXMLXML                   Type = "application/vnd.recordare.musicxml+xml"
-	ApplicationVNDRIMCod                                 Type = "application/vnd.rim.cod"
-	ApplicationVNDRNRealMedia                            Type = "application/vnd.rn-realmedia"
-	ApplicationVNDRoute66Link66XML                       Type = "application/vnd.route66.link66+xml"
-	ApplicationVNDSeemail                                Type = "application/vnd.seemail"
-	ApplicationVNDSEMA                                   Type = "application/vnd.sema"
-	ApplicationVNDSEMD                                   Type = "application/vnd.semd"
-	ApplicationVNDSEMf                                   Type = "application/vnd.semf"
-	ApplicationVNDShanaInformedFormdata                  Type = "application/vnd.shana.informed.formdata"
-	ApplicationVNDShanaInformedFormtemplate              Type = "application/vnd.shana.informed.formtemplate"
-	ApplicationVNDShanaInformedInterchange               Type = "application/vnd.shana.informed.interchange"
-	ApplicationVNDShanaInformedPackage                   Type = "application/vnd.shana.informed.package"
-	ApplicationVNDSimtechMindMapper                      Type = "application/vnd.simtech-mindmapper"
-	ApplicationVNSMAF                                    Type = "application/vnd.smaf"
-	ApplicationVNSmartTeacher                            Type = "application/vnd.smart.teacher"
-	ApplicationVNSolentSDKMXML                           Type = "application/vnd.solent.sdkm+xml"
-	ApplicationVNSPOTFIRE_DXP                            Type = "application/vnd.spotfire.dxp"
-	ApplicationVNSPOTFIRE_SFS                            Type = "application/vnd.spotfire.sfs"
-	ApplicationVNSQLite3                                 Type = "application/vnd.sqlite3"
-	ApplicationVNStarDivisionCalc                        Type = "application/vnd.stardivision.calc"
-	ApplicationVNStarDivisionDraw                        Type = "application/vnd.stardivision.draw"
-	ApplicationVNStarDivisionImpress                     Type = "application/vnd.stardivision.impress"
-	ApplicationVNStarDivisionMath                        Type = "application/vnd.stardivision.math"
-	ApplicationVNStarDivisionWriter                      Type = "application/vnd.stardivision.writer"
-	ApplicationVNStarDivisionWriterGlobal                Type = "application/vnd.stardivision.writer-global"
-	ApplicationVNSunXMLCalc                              Type = "application/vnd.sun.xml.calc"
-	ApplicationVNSunXMLCalcTemplate                      Type = "application/vnd.sun.xml.calc.template"
-	ApplicationVNSunXMLDraw                              Type = "application/vnd.sun.xml.draw"
-	ApplicationVNSunXMLDrawTemplate                      Type = "application/vnd.sun.xml.draw.template"
-	ApplicationVNSunXMLImpress                           Type = "application/vnd.sun.xml.impress"
-	ApplicationVNSunXMLImpressTemplate                   Type = "application/vnd.sun.xml.impress.template"
-	ApplicationVNSunXMLMath                              Type = "application/vnd.sun.xml.math"
-	ApplicationVNSunXMLWriter                            Type = "application/vnd.sun.xml.writer"
-	ApplicationVNSunXMLWriterGlobal                      Type = "application/vnd.sun.xml.writer.global"
-	ApplicationVNSunXMLWriterTemplate                    Type = "application/vnd.sun.xml.writer.template"
-	ApplicationVNSusCalendar                             Type = "application/vnd.sus-calendar"
-	ApplicationVNSVD                                     Type = "application/vnd.svd"
-	ApplicationVNSymbianInstall                          Type = "application/vnd.symbian.install"
-	ApplicationVNSyncMLXML                               Type = "application/vnd.syncml+xml"
-	ApplicationVNSyncMLDM_WBXML                          Type = "application/vnd.syncml.dm+wbxml"
-	ApplicationVNSyncMLDM_XML                            Type = "application/vnd.syncml.dm+xml"
-	ApplicationVNDTaoIntentModuleArchive                 Type = "application/vnd.tao.intent-module-archive"
-	ApplicationVNDTMobileLivetv                          Type = "application/vnd.tmobile-livetv"
-	ApplicationVNDTridTPT                                Type = "application/vnd.trid.tpt"
-	ApplicationVNDTriscapeMXS                            Type = "application/vnd.triscape.mxs"
-	ApplicationVNDTrueApp                                Type = "application/vnd.trueapp"
-	ApplicationVNUFDL                                    Type = "application/vnd.ufdl"
-	ApplicationVNUIQTheme                                Type = "application/vnd.uiq.theme"
-	ApplicationVNUmajin                                  Type = "application/vnd.umajin"
-	ApplicationVNUUnity                                  Type = "application/vnd.unity"
-	ApplicationVNUomlXML                                 Type = "application/vnd.uoml+xml"
-	ApplicationVNDVCX                                    Type = "application/vnd.vcx"
-	ApplicationVNDVisio                                  Type = "application/vnd.visio"
-	ApplicationVNDVisionary                              Type = "application/vnd.visionary"
-	ApplicationVNDVSF                                    Type = "application/vnd.vsf"
-	ApplicationVNDWapSIC                                 Type = "application/vnd.wap.sic"
-	ApplicationVNDWapSLC                                 Type = "application/vnd.wap.slc"
-	ApplicationVNDWapWBXML                               Type = "application/vnd.wap.wbxml"
-	ApplicationVNDWapWMLC                                Type = "application/vnd.wap.wmlc"
-	ApplicationVNDWapWMLScriptC                          Type = "application/vnd.wap.wmlscriptc"
-	ApplicationVNDWebTurbo                               Type = "application/vnd.webturbo"
-	ApplicationVNDWordPerfect                            Type = "application/vnd.wordperfect"
-	ApplicationVNDWQD                                    Type = "application/vnd.wqd"
-	ApplicationVNDWTSTF                                  Type = "application/vnd.wt.stf"
-	ApplicationVNDXara                                   Type = "application/vnd.xara"
-	ApplicationVNDXFDL                                   Type = "application/vnd.xfdl"
-	ApplicationVNDYamahaHVDic                            Type = "application/vnd.yamaha.hv-dic"
-	ApplicationVNDYamahaHVScript                         Type = "application/vnd.yamaha.hv-script"
-	ApplicationVNDYamahaHVVoice                          Type = "application/vnd.yamaha.hv-voice"
-	ApplicationVNDYamahaOpenScoreFormat                  Type = "application/vnd.yamaha.openscoreformat"
-	ApplicationVNDYamahaOpenScoreFormatOSFPVGXML         Type = "application/vnd.yamaha.openscoreformat.osfpvg+xml"
-	ApplicationVNDYamahaSMAFAudio                        Type = "application/vnd.yamaha.smaf-audio"
-	ApplicationVNDYamahaSMAFPhrase                       Type = "application/vnd.yamaha.smaf-phrase"
-	ApplicationVNDYellowRiverCustomMenu                  Type = "application/vnd.yellowriver-custom-menu"
-	ApplicationVNDZul                                    Type = "application/vnd.zul"
-	ApplicationVNDZZAZZDeckXML                           Type = "application/vnd.zzazz.deck+xml"
-	ApplicationVoiceXMLXML                               Type = "application/voicexml+xml"
-	ApplicationWASM                                      Type = "application/wasm"
-	ApplicationWinHLP                                    Type = "application/winhlp"
-	ApplicationWSDLXML                                   Type = "application/wsdl+xml"
-	ApplicationWSPolicyXML                               Type = "application/wspolicy+xml"
-	ApplicationX7ZCompressed                             Type = "application/x-7z-compressed"
-	ApplicationXAbiWord                                  Type = "application/x-abiword"
-	ApplicationXAceCompressed                            Type = "application/x-ace-compressed"
-	ApplicationXAuthorwareBin                            Type = "application/x-authorware-bin"
-	ApplicationXAuthorwareMap                            Type = "application/x-authorware-map"
-	ApplicationXAuthorwareSeg                            Type = "application/x-authorware-seg"
-	ApplicationXBCPIO                                    Type = "application/x-bcpio"
-	ApplicationXBittorrent                               Type = "application/x-bittorrent"
-	ApplicationXBzip                                     Type = "application/x-bzip"
-	ApplicationXBzip2                                    Type = "application/x-bzip2"
-	ApplicationXCDLink                                   Type = "application/x-cdlink"
-	ApplicationXChat                                     Type = "application/x-chat"
-	ApplicationXChessPGN                                 Type = "application/x-chess-pgn"
-	ApplicationXCPIO                                     Type = "application/x-cpio"
-	ApplicationXCSH                                      Type = "application/x-csh"
-	ApplicationXDebianPackage                            Type = "application/x-debian-package"
-	ApplicationXDirector                                 Type = "application/x-director"
-	ApplicationXDoom                                     Type = "application/x-doom"
-	ApplicationXDtbNcXXML                                Type = "application/x-dtbncx+xml"
-	ApplicationXDtbBookXML                               Type = "application/x-dtbook+xml"
-	ApplicationXDtbResourceXML                           Type = "application/x-dtbresource+xml"
-	ApplicationXDVI                                      Type = "application/x-dvi"
-	ApplicationXFontBDF                                  Type = "application/x-font-bdf"
-	ApplicationXFontGhostscript                          Type = "application/x-font-ghostscript"
-	ApplicationXFontLinuxPSF                             Type = "application/x-font-linux-psf"
-	ApplicationXFontOTF                                  Type = "application/x-font-otf"
-	ApplicationXFontPCF                                  Type = "application/x-font-pcf"
-	ApplicationXFontSNF                                  Type = "application/x-font-snf"
-	ApplicationXFontTTF                                  Type = "application/x-font-ttf"
-	ApplicationXFontType1                                Type = "application/x-font-type1"
-	ApplicationXFuturesplash                             Type = "application/x-futuresplash"
-	ApplicationXGnumeric                                 Type = "application/x-gnumeric"
-	ApplicationXGTAR                                     Type = "application/x-gtar"
-	ApplicationXGzip                                     Type = "application/x-gzip"
-	ApplicationXHDF                                      Type = "application/x-hdf"
-	ApplicationXISO9660Image                             Type = "application/x-iso9660-image"
-	ApplicationXJavaJNLPFile                             Type = "application/x-java-jnlp-file"
-	ApplicationXKIllustrator                             Type = "application/x-killustrator"
-	ApplicationXKrita                                    Type = "application/x-krita"
-	ApplicationXLatex                                    Type = "application/x-latex"
-	ApplicationXMobiPocketEbook                          Type = "application/x-mobipocket-ebook"
-	ApplicationXMSApplication                            Type = "application/x-ms-application"
-	ApplicationXMSWMD                                    Type = "application/x-ms-wmd"
-	ApplicationXMSWMZ                                    Type = "application/x-ms-wmz"
-	ApplicationXMSXBAP                                   Type = "application/x-ms-xbap"
-	ApplicationXMSAccess                                 Type = "application/x-msaccess"
-	ApplicationXMSBinder                                 Type = "application/x-msbinder"
-	ApplicationXMSCardfile                               Type = "application/x-mscardfile"
-	ApplicationXMSClip                                   Type = "application/x-msclip"
-	ApplicationXMSDownload                               Type = "application/x-msdownload"
-	ApplicationXMSMediaView                              Type = "application/x-msmediaview"
-	ApplicationXMSMetafile                               Type = "application/x-msmetafile"
-	ApplicationXMSMoney                                  Type = "application/x-msmoney"
-	ApplicationXMSPublisher                              Type = "application/x-mspublisher"
-	ApplicationXMSSchedule                               Type = "application/x-msschedule"
-	ApplicationXMSOpenOffice                             Type = "application/x-msterminal"
-	ApplicationXMSWrite                                  Type = "application/x-mswrite"
-	ApplicationXNetCDF                                   Type = "application/x-netcdf"
-	ApplicationXPerl                                     Type = "application/x-perl"
-	ApplicationXPKCS12                                   Type = "application/x-pkcs12"
-	ApplicationXPKCS7Certificates                        Type = "application/x-pkcs7-certificates"
-	ApplicationXPKCS7CertReqResp                         Type = "application/x-pkcs7-certreqresp"
-	ApplicationXPythonCode                               Type = "application/x-python-code"
-	ApplicationXRarCompressed                            Type = "application/x-rar-compressed"
-	ApplicationXRedHatPackageManager                     Type = "application/x-redhat-package-manager"
-	ApplicationXRPM                                      Type = "application/x-rpm"
-	ApplicationXSH                                       Type = "application/x-sh"
-	ApplicationXShar                                     Type = "application/x-shar"
-	ApplicationXShellscript                              Type = "application/x-shellscript"
-	ApplicationXShockwaveFlash                           Type = "application/x-shockwave-flash"
-	ApplicationXSilverlightApp                           Type = "application/x-silverlight-app"
-	ApplicationXSQLite3                                  Type = "application/x-sqlite3"
-	ApplicationXStuffit                                  Type = "application/x-stuffit"
-	ApplicationXStuffitX                                 Type = "application/x-stuffitx"
-	ApplicationXSV4CPIO                                  Type = "application/x-sv4cpio"
-	ApplicationXSV4CRC                                   Type = "application/x-sv4crc"
-	ApplicationXTAR                                      Type = "application/x-tar"
-	ApplicationXTCL                                      Type = "application/x-tcl"
-	ApplicationXTeX                                      Type = "application/x-tex"
-	ApplicationXTeXTFM                                   Type = "application/x-tex-tfm"
-	ApplicationXTeXInfo                                  Type = "application/x-texinfo"
-	ApplicationXTrash                                    Type = "application/x-trash"
-	ApplicationXUSTAR                                    Type = "application/x-ustar"
-	ApplicationXWaisSource                               Type = "application/x-wais-source"
-	ApplicationXX509CACert                               Type = "application/x-x509-ca-cert"
-	ApplicationXXFig                                     Type = "application/x-xfig"
-	ApplicationXXPIInstall                               Type = "application/x-xpinstall"
-	ApplicationXZipCompressed                            Type = "application/x-zip-compressed"
-	ApplicationXENCXML                                   Type = "application/xenc+xml"
-	ApplicationXHTMLXML                                  Type = "application/xhtml+xml"
-	ApplicationXML                                       Type = "application/xml"
-	ApplicationXMLDTD                                    Type = "application/xml-dtd"
-	ApplicationXOPXML                                    Type = "application/xop+xml"
-	ApplicationXSLTXML                                   Type = "application/xslt+xml"
-	ApplicationXSPFXML                                   Type = "application/xspf+xml"
-	ApplicationXVXML                                     Type = "application/xv+xml"
-	ApplicationYAML                                      Type = "application/yaml"
-	ApplicationZIP                                       Type = "application/zip"
-	ApplicationZIPCompressed                             Type = "application/zip-compressed"
-	ApplicationGolang                                    Type = "application/x-golang"
-	ApplicationJava                                      Type = "application/x-java"
-	// Audio Types
-	Audio3GPP2                  Type = "audio/3gpp2"
-	AudioAAC                    Type = "audio/aac"
-	AudioAACp                   Type = "audio/aacp"
-	AudioADPCM                  Type = "audio/adpcm"
-	AudioAIFF                   Type = "audio/aiff"
-	AudioBasic                  Type = "audio/basic"
-	AudioFLAC                   Type = "audio/flac"
-	AudioMIDI                   Type = "audio/midi"
-	AudioMP4                    Type = "audio/mp4"
-	AudioMPEG                   Type = "audio/mpeg"
-	AudioOGG                    Type = "audio/ogg"
-	AudioOpus                   Type = "audio/opus"
-	AudioVNDDigitalWinds        Type = "audio/vnd.digital-winds"
-	AudioVNDDTS                 Type = "audio/vnd.dts"
-	AudioVNDDTSHD               Type = "audio/vnd.dts.hd"
-	AudioVNDLucentVoice         Type = "audio/vnd.lucent.voice"
-	AudioVNDMSPlayReadyMediaPYA Type = "audio/vnd.ms-playready.media.pya"
-	AudioVNDNueraEcelp4800      Type = "audio/vnd.nuera.ecelp4800"
-	AudioVNDNueraEcelp7470      Type = "audio/vnd.nuera.ecelp7470"
-	AudioVNDNueraEcelp9600      Type = "audio/vnd.nuera.ecelp9600"
-	AudioVNDWAV                 Type = "audio/vnd.wav"
-	AudioWAV                    Type = "audio/wav"
-	AudioXWAV                   Type = "audio/x-wav"
-	AudioVNDWave                Type = "audio/vnd.wave"
-	AudioWebM                   Type = "audio/webm"
-	AudioXMatroska              Type = "audio/x-matroska"
-	AudioXMPEGURL               Type = "audio/x-mpegurl"
-	AudioXMSWAX                 Type = "audio/x-ms-wax"
-	AudioXMSWMA                 Type = "audio/x-ms-wma"
-	AudioXPNRealaudio           Type = "audio/x-pn-realaudio"
-	AudioXPNRealaudioPlugin     Type = "audio/x-pn-realaudio-plugin"
-	// Chemical Types
-	ChemicalXCDX  Type = "chemical/x-cdx"
-	ChemicalXCIF  Type = "chemical/x-cif"
-	ChemicalXCMDF Type = "chemical/x-cmdf"
-	ChemicalXCML  Type = "chemical/x-cml"
-	ChemicalXCSML Type = "chemical/x-csml"
-	ChemicalXYZ   Type = "chemical/x-xyz"
-	// Font Types
-	FontOTF   Type = "font/otf"
-	FontWOFF  Type = "font/woff"
-	FontWOFF2 Type = "font/woff2"
-	// Gcode Types
-	Gcode Type = "gcode"
-	// Image Types
-	ImageAVIF                  Type = "image/avif"
-	ImageBMP                   Type = "image/bmp"
-	ImageCGM                   Type = "image/cgm"
-	ImageG3FAX                 Type = "image/g3fax"
-	ImageGIF                   Type = "image/gif"
-	ImageHEIC                  Type = "image/heic"
-	ImageIEF                   Type = "image/ief"
-	ImageJPEG                  Type = "image/jpeg"
-	ImagePJPEG                 Type = "image/pjpeg"
-	ImagePNG                   Type = "image/png"
-	ImagePRSBTIF               Type = "image/prs.btif"
-	ImageSVGXML                Type = "image/svg+xml"
-	ImageTIFF                  Type = "image/tiff"
-	ImageVNDAdobePhotoshop     Type = "image/vnd.adobe.photoshop"
-	ImageVNDDJVU               Type = "image/vnd.djvu"
-	ImageVNDDWG                Type = "image/vnd.dwg"
-	ImageVNDDXF                Type = "image/vnd.dxf"
-	ImageVNDFastbidSheet       Type = "image/vnd.fastbidsheet"
-	ImageVNDFPX                Type = "image/vnd.fpx"
-	ImageVNDFST                Type = "image/vnd.fst"
-	ImageVNDFujixeroxEdmicsMMR Type = "image/vnd.fujixerox.edmics-mmr"
-	ImageVNDFujixeroxEdmicsRLC Type = "image/vnd.fujixerox.edmics-rlc"
-	ImageVNDMSMODI             Type = "image/vnd.ms-modi"
-	ImageVNDNetFPX             Type = "image/vnd.net-fpx"
-	ImageVNDCAPWAPWBMP         Type = "image/vnd.wap.wbmp"
-	ImageVNDXIFF               Type = "image/vnd.xiff"
-	ImageWEBP                  Type = "image/webp"
-	ImageXAdobeDNG             Type = "image/x-adobe-dng"
-	ImageXCanonCR2             Type = "image/x-canon-cr2"
-	ImageXCanonCRW             Type = "image/x-canon-crw"
-	ImageXCMURaster            Type = "image/x-cmu-raster"
-	ImageXCMX                  Type = "image/x-cmx"
-	ImageXEpsonERF             Type = "image/x-epson-erf"
-	ImageXFreehand             Type = "image/x-freehand"
-	ImageXFujiRAF              Type = "image/x-fuji-raf"
-	ImageXICNS                 Type = "image/x-icns"
-	ImageXICON                 Type = "image/x-icon"
-	ImageXKodakDCR             Type = "image/x-kodak-dcr"
-	ImageXKodakK25             Type = "image/x-kodak-k25"
-	ImageXKodakKDC             Type = "image/x-kodak-kdc"
-	ImageXMinoltaMRW           Type = "image/x-minolta-mrw"
-	ImageXNikonNEF             Type = "image/x-nikon-nef"
-	ImageXOlympusORF           Type = "image/x-olympus-orf"
-	ImageXPanasonicRaw         Type = "image/x-panasonic-raw"
-	ImageXPCX                  Type = "image/x-pcx"
-	ImageXPentaxPEF            Type = "image/x-pentax-pef"
-	ImageXPict                 Type = "image/x-pict"
-	ImageXPortableAnyMap       Type = "image/x-portable-anymap"
-	ImageXPortableBitmap       Type = "image/x-portable-bitmap"
-	ImageXPortableGrayMap      Type = "image/x-portable-graymap"
-	ImageXPortablePixmap       Type = "image/x-portable-pixmap"
-	ImageXRGB                  Type = "image/x-rgb"
-	ImageXSigmaX3F             Type = "image/x-sigma-x3f"
-	ImageXSonyARW              Type = "image/x-sony-arw"
-	ImageXSonySR2              Type = "image/x-sony-sr2"
-	ImageXSonySRF              Type = "image/x-sony-srf"
-	ImageXXBitmap              Type = "image/x-xbitmap"
-	ImageXXPixMap              Type = "image/x-xpixmap"
-	ImageXXWindowDump          Type = "image/x-xwindowdump"
-	// Message Types
-	MessageRFC822 Type = "message/rfc822"
-	// Model Types
-	ModelIGES   Type = "model/iges"
-	ModelMesh   Type = "model/mesh"
-	ModelVNDWF  Type = "model/vnd.dwf"
-	ModelVNDGDL Type = "model/vnd.gdl"
-	ModelVNDGTW Type = "model/vnd.gtw"
-	ModelVNDMTS Type = "model/vnd.mts"
-	ModelVNDVTU Type = "model/vnd.vtu"
-	ModelVRML   Type = "model/vrml"
-	// Test Types
-	TestMIMEType     Type = "test/mimetype"
-	TestMIMETypeTest Type = "test/mimetype/test"
-	// Text Types
-	TextCalendar           Type = "text/calendar"
-	TextCSS                Type = "text/css"
-	TextCSV                Type = "text/csv"
-	TextHTML               Type = "text/html"
-	TextJavaScript         Type = "text/javascript"
-	TextMarkdown           Type = "text/markdown"
-	TextMathML             Type = "text/mathml"
-	TextPlain              Type = "text/plain"
-	TextPRSLinesTag        Type = "text/prs.lines.tag"
-	TextRichText           Type = "text/richtext"
-	TextSGML               Type = "text/sgml"
-	TextTabSeparatedValues Type = "text/tab-separated-values"
-	TextTROFF              Type = "text/troff"
-	TextURIList            Type = "text/uri-list"
-	TextVNDWAPSI           Type = "text/vnd.wap.si"
-	TextVNDWAPSL           Type = "text/vnd.wap.sl"
-	TextVNDWAPWML          Type = "text/vnd.wap.wml"
-	TextVNDWAPWMLScript    Type = "text/vnd.wap.wmlscript"
-	TextXASM               Type = "text/x-asm"
-	TextXC                 Type = "text/x-c"
-	TextXFortran           Type = "text/x-fortran"
-	TextXJavaSource        Type = "text/x-java-source"
-	TextXPascal            Type = "text/x-pascal"
-	TextXPython            Type = "text/x-python"
-	TextXSetext            Type = "text/x-setext"
-	TextXUUEncode          Type = "text/x-uuencode"
-	TextXVCalendar         Type = "text/x-vcalendar"
-	TextXVCard             Type = "text/x-vcard"
-	// Video Types
-	Video3GPP                   Type = "video/3gpp"
-	Video3GPP2                  Type = "video/3gpp2"
-	VideoH261                   Type = "video/h261"
-	VideoH263                   Type = "video/h263"
-	VideoH264                   Type = "video/h264"
-	VideoJPEG                   Type = "video/jpeg"
-	VideoJPM                    Type = "video/jpm"
-	VideoMJ2                    Type = "video/mj2"
-	VideoMP2T                   Type = "video/mp2t"
-	VideoMP4                    Type = "video/mp4"
-	VideoMPEG                   Type = "video/mpeg"
-	VideoOGG                    Type = "video/ogg"
-	VideoQuickTime              Type = "video/quicktime"
-	VideoVNDFVT                 Type = "video/vnd.fvt"
-	VideoVNDPUGURL              Type = "video/vnd.mpegurl"
-	VideoVNDMSPlayReadyMediaPYV Type = "video/vnd.ms-playready.media.pyv"
-	VideoVNDUVIVO               Type = "video/vnd.vivo"
-	VideoWEBM                   Type = "video/webm"
-	VideoXF4V                   Type = "video/x-f4v"
-	VideoXFli                   Type = "video/x-fli"
-	VideoXFLV                   Type = "video/x-flv"
-	VideoXM4V                   Type = "video/x-m4v"
-	VideoXMatroska              Type = "video/x-matroska"
-	VideoXMSASF                 Type = "video/x-ms-asf"
-	VideoXMSWM                  Type = "video/x-ms-wm"
-	VideoXMSWMV                 Type = "video/x-ms-wmv"
-	VideoXMSWMX                 Type = "video/x-ms-wmx"
-	VideoXMSWVX                 Type = "video/x-ms-wvx"
-	VideoXMSVideo               Type = "video/x-msvideo"
-	VideoXSGIMovie              Type = "video/x-sgi-movie"
-	// Conference Types
-	XConferenceXCooltalk Type = "x-conference/x-cooltalk"
-)
